avcodec/lossless_videodsp: use ptrdiff_t for length parameters

+++ libavcodec/x86/lossless_videodsp_init.c
@@ -23,27 +23,27 @@
 #include "../lossless_videodsp.h"
 #include "libavutil/x86/cpu.h"
 
-void ff_add_bytes_mmx(uint8_t *dst, uint8_t *src, intptr_t w);
-void ff_add_bytes_sse2(uint8_t *dst, uint8_t *src, intptr_t w);
+void ff_add_bytes_mmx(uint8_t *dst, uint8_t *src, ptrdiff_t w);
+void ff_add_bytes_sse2(uint8_t *dst, uint8_t *src, ptrdiff_t w);
 
 void ff_add_median_pred_mmxext(uint8_t *dst, const uint8_t *top,
-                               const uint8_t *diff, intptr_t w,
+                               const uint8_t *diff, ptrdiff_t w,
                                int *left, int *left_top);
 void ff_add_median_pred_sse2(uint8_t *dst, const uint8_t *top,
-                             const uint8_t *diff, intptr_t w,
+                             const uint8_t *diff, ptrdiff_t w,
                              int *left, int *left_top);
 
 int  ff_add_left_pred_ssse3(uint8_t *dst, const uint8_t *src,
-                            intptr_t w, int left);
+                            ptrdiff_t w, int left);
 int  ff_add_left_pred_unaligned_ssse3(uint8_t *dst, const uint8_t *src,
-                                      intptr_t w, int left);
+                                      ptrdiff_t w, int left);
 
-int ff_add_left_pred_int16_ssse3(uint16_t *dst, const uint16_t *src, unsigned mask, int w, unsigned acc);
-int ff_add_left_pred_int16_sse4(uint16_t *dst, const uint16_t *src, unsigned mask, int w, unsigned acc);
+int ff_add_left_pred_int16_ssse3(uint16_t *dst, const uint16_t *src, unsigned mask, ptrdiff_t w, unsigned acc);
+int ff_add_left_pred_int16_sse4(uint16_t *dst, const uint16_t *src, unsigned mask, ptrdiff_t w, unsigned acc);
 
 #if HAVE_INLINE_ASM && HAVE_7REGS && ARCH_X86_32
 static void add_median_pred_cmov(uint8_t *dst, const uint8_t *top,
-                                 const uint8_t *diff, intptr_t w,
+                                 const uint8_t *diff, ptrdiff_t w,
                                  int *left, int *left_top)
 {
     x86_reg w2 = -w;
